requirements.txt
awsebcli==3.20.0
botocore==1.20.108
cement==2.8.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.3
Flask==2.0.1
future==0.16.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
numpy==1.21.0
pandas==1.3.0
pathspec==0.5.9
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
semantic-version==2.8.5
six==1.14.0
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
Werkzeug==2.0.1
